refactor handlers under workflow_run
