--- api/src/handlers/stop.js
@@ -1,7 +1,0 @@
-import { readPinnedRunId } from './utils.js';
-
-export function createStopHandler( client ) {
-  return async ( req, res ) => {
-    res.json( await client.stopWorkflow( req.params.id, readPinnedRunId( req ) ) );
-  };
-}
